Update method implemented.
